+++ bin/aecli-name.js
@@ -32,7 +32,7 @@ program
   .option('-u, --url [hostname]', 'Node to connect to', utils.constant.EPOCH_URL)
   .option('-U, --internalUrl [internal]', 'Node to connect to(internal)', utils.constant.EPOCH_INTERNAL_URL)
   .option('--networkId [networkId]', 'Network id (default: ae_mainnet)')
-  .option('--native', 'Build transaction natively', true)
+  .option('--native', 'Build transaction natively')
   .option('-P, --password [password]', 'Wallet Password')
   .option('-N, --nameTtl [nameTtl]', 'Name life Ttl', utils.constant.NAME_TTL)
   .option('-T, --ttl [ttl]', 'Life Ttl', utils.constant.AENS_TX_TTL)