+++ bin/aecli.js
@@ -33,6 +33,7 @@ const EXECUTABLE_CMD = [
   { name: 'account', desc: 'Handle wallet operations' },
   { name: 'contract', desc: 'Compile contracts' },
   { name: 'name', desc: 'AENS system' },
+  { name: 'tx', desc: 'Transaction builder' },
   /* TODO implement oracle module
   * {name: 'oracle', desc: 'Interact with oracles'},
   */