+++ bin/utils/constant.js
@@ -29,9 +29,9 @@ export const HASH_TYPES = {
 }
 
 // ## CONNECTION
-export const EPOCH_URL = 'https://sdk-edgenet.aepps.com'
-export const EPOCH_INTERNAL_URL = 'https://sdk-edgenet.aepps.com'
-export const EPOCH_WEBSOCKET_URL = 'https://sdk-edgenet.aepps.com'
+export const EPOCH_URL = 'https://sdk-mainnet.aepp.com'
+export const EPOCH_INTERNAL_URL = 'https://sdk-mainnet.aepp.com'
+export const EPOCH_WEBSOCKET_URL = 'https://sdk-testnet.aepps.com'
 
 // ## CHAIN
 export const PLAY_LIMIT = 10
@@ -44,6 +44,8 @@ export const GAS = 1600000 - 21000 // MAX GAS
 // ## AENS
 export const AENS_TX_TTL = 500
 export const NAME_TTL = 50000
+export const CLIENT_TTL = 1
 
 // ## ACCOUNT
 export const SPEND_TX_TTL = 500
+export const SPEND_TX_FEE = 16740