+++ bin/utils/helpers.js
@@ -112,3 +112,27 @@ export function readFile (path, encoding = null, errTitle = 'READ FILE ERR') {
     }
   }
 }
+
+// ## AENS helpers methods
+
+// Get `name` status
+export function updateNameStatus (name) {
+ return async (client) => {
+   try {
+     return await client.api.getNameEntryByName(name)
+   } catch (e) {
+     if (e.response && e.response.status === 404) {
+       return {name, status: 'AVAILABLE'}
+     }
+     throw e
+   }
+ }
+}
+
+// Check if `name` is `AVAILABLE`
+export function isAvailable (name) { return name.status === 'AVAILABLE' }
+
+// Validate `name`
+export function validateName (name) {
+  if (R.last(name.split('.')) !== 'test') { throw new Error('AENS TLDs must end in .test') }
+}